% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkProcessing.R
\name{BulkProcessing}
\alias{BulkProcessing}
\title{Processing and analyzing Multiple Experiments}
\usage{
BulkProcessing(data, do_analysis = TRUE, params = list(), verbose = FALSE)
}
\arguments{
\item{data}{Compiled data of experiments.}

\item{do_analysis}{Logical. Whether statistic analysis is included. Default is \code{TRUE}.}

\item{params}{A list of parameters for various processing steps. If a parameter is not
provided, default values will be used. \code{control} used in \code{GetAnalysis()} is required.
parameters can be provided for the following functions:
\itemize{
\item \code{ConvertTime}
\item \code{CleanMeta}
\item \code{CleanRaw}
\item \code{GetCalculation}
\item \code{SpreadCalculation}
\item \code{GetAnalysis}
\item \code{SummarizeResult}
}}

\item{verbose}{Logical; if TRUE, prints detailed processing information for troubleshooting.
Default is FALSE.}
}
\value{
A list containing three elements:
\itemize{
\item combined_calculation: A data frame of combined calculations from all experiments
\item combined_cleanraw: Cleaned raw data for each experiment
\item combined_result: A data frame of combined results from all experiments
}
}
\description{
This function processes multiple experiments in parallel, performing a series of operations
including time conversion, metadata cleaning, raw data cleaning, calculations, analysis,
and result summarization.
}
\examples{
if (interactive()) {

#Set parameters that will be used in the process

myparams = list(
CleanMeta = list(split_content = TRUE, split_into = c("dilution", "sampleID")),
GetAnalysis = list(control = 'neg', alternative = "greater"),
SummarizeResult = list(sig_method = 'metric_count', method_threshold = 3)
)

results = ParallelProcessing(data = my_data, params = my_params)

# Access combined results
combined_results <- results$combined_result

# Access cleaned raw data for a specific plate
cleaned_raw_data_plate1 <- results$combined_cleanraw[[1]]
}

}
