% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BulkReading.R
\name{BulkReading}
\alias{BulkReading}
\title{Read MARS data from multiple folders}
\usage{
BulkReading(path, plate_subfix, raw_subfix, helper_func = NULL)
}
\arguments{
\item{path}{Character string specifying the path to the parent directory containing the data folders.}

\item{plate_subfix}{Character string used to identify plate data files.}

\item{raw_subfix}{Character string used to identify raw data files.}

\item{helper_func}{Optional function to be applied to each column of the plate data (default: NULL).}
}
\value{
A list containing data from each folder, including plate, raw, and replicate data.
}
\description{
Read MARS data from multiple folders
}
\examples{
\dontrun{
grinder_data <- ReadMARS(path = './data/grinder/',
                         plate_subfix = 'plate',
                         raw_subfix = 'raw',
                         helper_func = flip_and_replace)
}
}
