% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanRaw.R
\name{CleanRaw}
\alias{CleanRaw}
\title{Generate Clean Raw Data}
\usage{
CleanRaw(meta, raw, plate_time, cycle_total)
}
\arguments{
\item{meta}{A data frame containing the metadata. Output from CleanMeta function.}

\item{raw}{Raw fluorescence readings from MARS software.}

\item{plate_time}{Output of `ConvertTime()`.}

\item{cycle_total}{The total number of cycles (rows) to include in the output.}
}
\value{
A data frame containing the cleaned raw fluorescence data.
}
\description{
This function takes metadata, raw data, and total cycle information to generate clean raw fluorescence data.
}
\examples{
\dontrun{
meta <- data.frame(content_replicate = c("A1", "A2", "A3"))
raw <- data.frame(
  X1 = c("Header", "1", "2", "3"),
  X2 = c("Time", "0", "1", "2"),
  A1 = c("Value", "100", "110", "120"),
  A2 = c("Value", "200", "210", "220"),
  A3 = c("Value", "300", "310", "320")
)
plate_time <- c(0, 1, 2)

cleaned_data <- CleanRaw(meta, raw, plate_time)
print(cleaned_data)
}
}
