% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetAnalysis.R
\name{GetAnalysis}
\alias{GetAnalysis}
\title{Perform Statistical Analysis on Calculations}
\usage{
GetAnalysis(
  calculation_spread,
  control,
  test = "wilcox",
  alternative = "two.sided",
  adjust_p = FALSE,
  alpha = 0.05
)
}
\arguments{
\item{calculation_spread}{A list of data frames, each representing a calculation spread. Output of \code{SpreadCaculation()}.}

\item{control}{The name or pattern of the control column in each data frame.}

\item{test}{The statistical test to use. Options are "t-test", "wilcox", or "yuen". Default is "wilcox".}

\item{alternative}{Options are "two.sided", "less", or "greater". Default is "two.sided".}

\item{adjust_p}{Logical. Whether to adjust p-values for multiple comparisons. Default is FALSE.}

\item{alpha}{The significance level for determining significance stars. Default is 0.05.}
}
\value{
A list of data frames containing the results of the statistical analysis for selected metrics.
}
\description{
This function performs statistical analysis on a list of calculation spreads,
comparing each column to a control column using various statistical tests.
}
\examples{
\dontrun{
analysis <- GetAnalysis(calculation_spread, control = "neg", test = "wilcox", tail = 'greater')
}

}
\seealso{
\code{\link[stats]{t.test}}, \code{\link[stats]{wilcox.test}}, \code{\link[WRS2]{yuen}}
}
