% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{null-default}
\alias{null-default}
\alias{\%||\%}
\alias{This}
\alias{operator}
\alias{returns}
\alias{the}
\alias{left-hand}
\alias{side}
\alias{if}
\alias{it's}
\alias{not}
\alias{NULL,}
\alias{and}
\alias{right-hand}
\alias{otherwise.}
\title{Null default operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{The primary value}

\item{y}{The default value to be used if x is NULL}
}
\value{
x if it's not NULL, otherwise y
}
\description{
Null default operator
}
\examples{
NULL \%||\% 1  # Returns 1
5 \%||\% 1     # Returns 5

}
