% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetric.R
\name{PlotMetric}
\alias{PlotMetric}
\title{Create a customizable ggplot}
\usage{
PlotMetric(
  calculation,
  x = "content",
  y = "RAF",
  point = TRUE,
  box = TRUE,
  base_theme = theme_bw(),
  ...
)
}
\arguments{
\item{calculation}{A data frame containing the data to be plotted. Output of `GetCalculation`}

\item{x}{Character string specifying the column name for the x-axis. Default is "content".}

\item{y}{Character string specifying the column name for the y-axis. Default is "RAF".}

\item{point}{Logical. If TRUE (default), adds jittered points to the plot.}

\item{box}{Logical. If TRUE (default), adds a boxplot to the plot.}

\item{base_theme}{A ggplot2 theme object. Default is theme_bw().}

\item{...}{Additional ggplot2 functions to be applied to the plot.}
}
\value{
A ggplot object.
}
\description{
This function generates a ggplot object with options for boxplot and jittered points. 
Additional ggplot2 functions to be applied to the plot.
}
\examples{
\dontrun{
PlotMetric(calculation, x = "content", y = "MS",
           labs(title = "Max Slope by Sample"),
           scale_y_continuous(limits = c(0, 1600)))
}
}
