% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPlate.R
\name{PlotPlate}
\alias{PlotPlate}
\title{Plot Time Series Data}
\usage{
PlotPlate(raw, plate_time, format = 96, f_size = 5, fill = FALSE)
}
\arguments{
\item{raw}{A data frame containing the raw plate data. The first row and
first two columns are assumed to be metadata and are removed.}

\item{plate_time}{Output from \code{ConvertTime()}.}

\item{format}{Format of plates used in the experiment. 96 or 384.}

\item{f_size}{font size for subtitles.}

\item{fill}{Logical, whether to fill in missing wells with 0. Default is FALSE.}
}
\value{
A ggplot object representing the plate data.
}
\description{
This function creates a faceted plot of time series data for each well in a plate layout.
}
\examples{
if (interactive()) {
PlotPlate(raw, plate_time)
}

}
