% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRawMulti.R
\name{PlotRawMulti}
\alias{PlotRawMulti}
\title{Plot Multiple Samples from the Cleaned Raw Data}
\usage{
PlotRawMulti(
  raw,
  samples,
  legend_position = "topleft",
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{raw}{Cleaned raw data. Output from \code{GetCleanRaw}.}

\item{samples}{The names of the samples to plot.}

\item{legend_position}{Position of legend. Default is "topleft".
Choose from "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{ylim}{Numeric vector giving the y coordinate range (relative fluorescence units). If NULL (default), limits are computed from the data.}

\item{xlim}{Numeric vector giving the x coordinate range (hours). If NULL (default), limits are computed from the data.}
}
\description{
The function visualizes the fluorescence over time for selected samples.
}
\examples{
\dontrun{
PlotRawMulti(raw = raw, samples = samples)
}
}
