% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCalculation.R
\name{GetCalculation}
\alias{GetCalculation}
\title{Perform Calculations}
\usage{
GetCalculation(
  raw,
  meta,
  norm = FALSE,
  norm_ct,
  threshold_method = "stdv",
  time_skip = 5,
  sd_fold = 3,
  bg_fold = 3,
  cycle_background = 4,
  binw = 6
)
}
\arguments{
\item{raw}{Cleaned raw data matrix. Output from `CleanRaw()`.}

\item{meta}{Cleaned meta data. Output from `CleanMeta()`.}

\item{norm}{Logical. If TRUE, normalization will be performed. Default is FALSE.}

\item{norm_ct}{Sample name used to normalize calculation.}

\item{threshold_method}{Method for calculating threshold ('stdv' or 'bg_ratio').}

\item{time_skip}{Number of initial time points to skip when checking for threshold crossing.
This helps ignore early crossings that may be due to reasons unrelated to seeding activity.}

\item{sd_fold}{Fold of standard deviation to calculate the threshold for RAF (for 'stdv' method).}

\item{bg_fold}{Background fold for threshold calculation (for 'bg_ratio' method).}

\item{cycle_background}{The cycle number chosen as the background for RAF calculations.}

\item{binw}{Bin width for the MS calculation.}
}
\value{
A data frame containing the results of the calculation.
}
\description{
This function takes cleaned raw data and performs various analyses,
including calculating the time to threshold, Rate of Amyloid Formation (RAF), Max Point Ratio (MPR), 
Max Slope (MS), and whether the reaction crosses the threshold (XTH).
}
\examples{
\dontrun{
GetCalculation(raw, meta, norm = TRUE, norm_ct = 'pos', 'stdv')
}
}
