% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotMetric.R
\name{PlotMetric}
\alias{PlotMetric}
\title{Create a customizable ggplot for metrics}
\usage{
PlotMetric(
  calculation,
  x = "content",
  y = "RAF",
  fill_var = NULL,
  point = TRUE,
  box = TRUE,
  base_theme = theme_bw(),
  ...
)
}
\arguments{
\item{calculation}{A data frame containing the data to be plotted. Output of \code{GetCalculation}}

\item{x}{Character string specifying the column name for the x-axis. Default is "content".}

\item{y}{Character string specifying the column name for the y-axis. Default is "RAF".}

\item{fill_var}{The name of the column to be used for fill color, or a vector of colors.
If NULL, no fill color is applied. Default is NULL.}

\item{point}{Logical. If TRUE (default), adds jittered points to the plot.}

\item{box}{Logical. If TRUE (default), adds a boxplot to the plot.}

\item{base_theme}{A ggplot2 theme object. Default is theme_bw().}

\item{...}{Additional ggplot2 functions to be applied to the plot.}
}
\value{
A ggplot object.
}
\description{
This is wrapper function to generate a ggplot object with default options for boxplot and jittered points.
Additional ggplot2 functions to be applied to the plot.
}
\examples{
\dontrun{
PlotMetric(calculation_96, y = "MS", point = FALSE, box = FALSE,
           boxplot= geom_boxplot(color = 'gray'),
           scatter = geom_point(color = "blue") ,
           xlab = xlab("Sample ID"),
           ylab =ylab("Normalized Max Slope"))
}
}
