% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRawMulti.R
\name{PlotRawMulti}
\alias{PlotRawMulti}
\title{Plot Multiple Samples from the Cleaned Raw Data}
\usage{
PlotRawMulti(
  raw,
  samples,
  legend_position = "topleft",
  xlim = NULL,
  ylim = NULL,
  custom_colors = NULL,
  xlab = "Time (h)",
  ylab = "Fluorescence",
  linetype = "l"
)
}
\arguments{
\item{raw}{Cleaned raw data. Output from \code{GetCleanRaw}.}

\item{samples}{The names of the samples to plot.}

\item{legend_position}{Position of legend. Default is "topleft".
Choose from "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}

\item{xlim}{Numeric vector giving the x coordinate range (hours). If NULL (default), limits are computed from the data.}

\item{ylim}{Numeric vector giving the y coordinate range (relative fluorescence units). If NULL (default), limits are computed from the data.}

\item{custom_colors}{An optional vector of colors to be used for plotting.
If NULL (default), the function will use the original color scheme.}

\item{xlab}{Label for x-axis.}

\item{ylab}{Label for y-axis.}

\item{linetype}{The type parameter in plotting functions accepts various values to control the appearance of the plot: "p" draws only points,
"l" creates lines, "b" combines both points and lines, "c" displays empty points joined by lines, "o" overplots points and lines, "s" and "S" create stair steps,
"h" produces histogram-like vertical lines, and "n" sets up the plot without drawing any points or lines. For more detailed information on these options,
refer to the documentation of the plot function.}
}
\description{
The function visualizes the fluorescence over time for selected samples.
}
\examples{
if (interactive()) {
PlotRawMulti(raw = raw, samples = samples)
}
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988) The New S Language. Wadsworth & Brooks/Cole.
}
