% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotRawSingle.R
\name{PlotRawSingle}
\alias{PlotRawSingle}
\title{Plot a Single Sample from the Cleaned Raw Data}
\usage{
PlotRawSingle(
  raw,
  sample,
  legend_position = "bottomright",
  ylim = NULL,
  xlim = NULL
)
}
\arguments{
\item{raw}{Cleaned raw data. Output from \code{GetCleanRaw}}

\item{sample}{The name of the sample to plot.}

\item{legend_position}{Position of legend. Default is "topleft".}

\item{ylim}{Numeric vector giving the y coordinate range (relative fluorescence units). If NULL (default), limits are computed from the data.}

\item{xlim}{Numeric vector giving the x coordinate range (hours). If NULL (default), limits are computed from the data.
Choose from "bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright", "right" and "center".}
}
\description{
The function plots fluorescence for selected sample over time.
}
\examples{
\dontrun{
PlotRawSingle(raw = raw, sample = sample)
}
}
