% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarizeResult.R
\name{SummarizeResult}
\alias{SummarizeResult}
\title{Summarize Analysis Results}
\usage{
SummarizeResult(
  analysis,
  calculation,
  sig_method = "metric_count",
  method_threshold = 3
)
}
\arguments{
\item{analysis}{Output of \code{GetAnalysis()}.}

\item{calculation}{Output of \code{GetCalculation()}.}

\item{sig_method}{Method to determine overall significance.
Options are "metric_count", "xth_count", or any of the names in the analysis list. Default is "metric_count".}

\item{method_threshold}{Threshold for count when using "metric_count" and "xth_count" method. Default is 3.}
}
\value{
A data frame summarizing results of the analysis and calculation, with columns:

\itemize{
\item result: Overall result based on the chosen method

\item position: Location of the sample replications based on plate layout

\item method: Method chosen for determination of the result

\item *_sig: Significance indicator for each metric

\item *_p: P-value or adjusted p-value for each metric

\item metric_count: Count of significant results across all analyses

\item xth_count: The number of reactions crossing the fluorescent threshold

\item total_rep: Total number of replicates

\item xth_percent: Percentage of replicates crossing the fluorescent threshold
}
}
\description{
This function combines analysis results from multiple tests with metadata,
and determines overall significance based on a specified method.
}
\examples{
\dontrun{
result <- SummarizeResult(analysis, calculation)
result_mpr <- SummarizeResult(analysis, calculation, sig_method = "MPR")
}

}
